backend/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
email-validator
pymongo[srv]
python-multipart
cloudinary
requests